Afegit multiples eines 1. Controlador d'excepcions 2. Correcció dels errors de compatibilitat amb les carpetes 3. Afegit la possibilitat de fer el readAll

+++ src/controlador/ProvaEscalador.java
@@ -1,17 +1,23 @@
 package controlador;
 
 import dao.DBConnection;
-import dao.mysql.SQLiteEscaladorDAO;
+import dao.SQLite.SQLiteEscaladorDAO;
 import model.Escalador;
 import personalExceptions.*;
 
 import java.sql.Connection;
+import java.sql.PreparedStatement;
+import java.sql.ResultSet;
+import java.sql.SQLException;
 import java.util.Scanner;
+
+import static dao.SQLite.SQLiteViaDAO.connection;
+
 
 public class ProvaEscalador {
     private static Scanner scanner = new Scanner(System.in);
     public static void main(String[] args) {
-        String url = "jdbc:sqlite:..\\..\\db\\vies_db2.db";
+        String url = "jdbc:sqlite:..\\..\\db\\vies_db1.db";
         Connection connection = null;
 
         try {
@@ -24,10 +30,11 @@ public class ProvaEscalador {
             do {
                 // Display menu
                 System.out.println("\n--- Menu Escalador ---");
-                System.out.println("1. Crear Escalador");
+                System.out.println("1. Insertar Escalador");
                 System.out.println("2. Leer Escalador");
                 System.out.println("3. Actualizar Escalador");
                 System.out.println("4. Eliminar Escalador");
+                System.out.println("5. Mostrar todos los Escaladores");
                 System.out.println("5. Salir");
                 System.out.print("Elige una opción: ");
                 option = scanner.nextInt();
@@ -35,32 +42,34 @@ public class ProvaEscalador {
                 switch (option) {
                     case 1:
                         try {
-                            System.out.println("--- Crear Escalador ---");
-                            // Create Escalador
+                            System.out.println("--- Insertar Escalador ---");
+                            // Comprovem que el ID no sigui negatiu
                             System.out.print("ID: ");
-                            int id = scanner.nextInt();
-                            scanner.nextLine(); // Consume newline
+                            int id = insertID();
+                            // Comprovem que el nom no sigui buit
                             System.out.print("Nombre: ");
-                            String nombre = scanner.nextLine();
+                            String nombre = insertName();
+                            // Comprovem que l'alias no sigui buit
                             System.out.print("Alias: ");
-                            String alias = scanner.nextLine();
+                            String alias = insertAlias();
+                            // Comprovem que l'edat no sigui negativa i no sigui superior a 120
                             System.out.print("Edad: ");
-                            int edad = scanner.nextInt();
-                            scanner.nextLine(); // Consume newline
+                            int edad = insertAge();
+                            // Comprovem que el nivell sigui correcte
                             System.out.print("Nivel Máximo: ");
-                            String nivelMax = scanner.nextLine();
-                            String estilo = null;
+                            String nivelMax = insertNivellMaxim();
+                            // Comprovem si el nivell es correcte
                             System.out.print("Estilo Preferido: ");
-                            estilo = scanner.nextLine();
-                            if (!estilo.equalsIgnoreCase("esportiva") && !estilo.equalsIgnoreCase("clàssica") && !estilo.equalsIgnoreCase("gel")) {
-                                throw new EstilPreferitException("Estil preferit no vàlid, ha de ser; esportiva | clàssica | gel");
-                            }
+                            String estilo = insertEstilPreferit();
+                            // No comprovem res de la fita
                             System.out.print("Fita: ");
                             String fita = scanner.nextLine();
+                            // Comprovem si la via existeix
                             System.out.print("ID Via Máxima: ");
-                            int idViaMax = scanner.nextInt();
+                            int idViaMax = insertViaMax();
+                            // Una vegada comprovat totes les dades es pot inserir l'Escalador
                             Escalador nuevoEscalador = new Escalador(id, nombre, alias, edad, nivelMax, estilo, fita, idViaMax);
-                            escaladorDAO.createTable(nuevoEscalador);
+                            escaladorDAO.insertTable(nuevoEscalador);
                             System.out.println("Escalador creado con éxito.");
                         } catch (EstilPreferitException e) {
                             System.err.println(e.getMessage());
@@ -69,53 +78,71 @@ public class ProvaEscalador {
 
                     case 2:
                         System.out.println("--- Leer Escalador ---");
-                        // Read Escalador
+                        // Comprovem si l'escalador existeix abans d'intentar llegir
                         System.out.print("ID del escalador a leer | ");
                         int idLeer = scanner.nextInt();
+                        if (!doesEscaladorExist(idLeer)) {
+                            System.err.println("El escalador no existe.");
+                            break;
+                        } else {
                         escaladorDAO.readTable(idLeer);
+                        }
                         break;
-
                     case 3:
                         try {
                             System.out.println("--- Actualizar Escalador ---");
-                            // Update Escalador
+                            // Comprovem que el ID no sigui negatiu
                             System.out.print("ID del escalador a actualizar: ");
-                            int idActualizar = scanner.nextInt();
-                            scanner.nextLine(); // Consume newline
+                            int idActualizar = insertID();
+                            // Comprovem que el nom no sigui buit
                             System.out.print("Nuevo Nombre: ");
-                            String nuevoNombre = scanner.nextLine();
+                            String nuevoNombre = insertName();
+                            // Comprovem que l'alias no sigui buit
                             System.out.print("Nuevo Alias: ");
-                            String nuevoAlias = scanner.nextLine();
+                            String nuevoAlias = insertAlias();
+                            // Comprovem que l'edat no sigui negativa i no sigui superior a 120
                             System.out.print("Nueva Edad: ");
-                            int nuevaEdad = scanner.nextInt();
-                            scanner.nextLine(); // Consume newline
+                            int nuevaEdad = insertAge();
+                            // Comprovem que el nivell sigui correcte
                             System.out.print("Nuevo Nivel Máximo: ");
-                            String nuevoNivelMax = scanner.nextLine();
-                            // S'ha de fer en format funcions personals com aquestes
-                            String nuevoEstilo = UptadeEstilPreferit();
-                            // D'aquesta forma el try/catch final no es tan llarg
+                            String nuevoNivelMax = insertNivellMaxim();
+                            // Comprovem si el nivell es correcte
+                            System.out.print("Nuevo Estilo Preferido: ");
+                            String nuevoEstilo = insertEstilPreferit();
+                            // No comprovem res de la fita
                             System.out.print("Nueva Fita: ");
                             String nuevaFita = scanner.nextLine();
+                            // Comprovem si la via existeix
                             System.out.print("Nuevo ID Via Máxima: ");
-                            int nuevoIdViaMax = scanner.nextInt();
+                            int nuevoIdViaMax = insertViaMax();
+                            // Una vegada tot esta comporvat es pot actualitzar l'escalador
                             Escalador escaladorActualizado = new Escalador(idActualizar, nuevoNombre, nuevoAlias, nuevaEdad, nuevoNivelMax, nuevoEstilo, nuevaFita, nuevoIdViaMax);
                             escaladorDAO.updateTable(escaladorActualizado);
                             System.out.println("Escalador actualizado con éxito.");
-                        } catch (personalExceptions.EstilPreferitException e) {
+                        } catch (EstilPreferitException e) {
                             System.err.println("Error: " + e.getMessage());
                         }
                         break;
 
                     case 4:
                         System.out.println("--- Eliminar Escalador ---");
-                        // Delete Escalador
+                        // Comprovem que l'escalador existeix abans d'eliminarlo
                         System.out.print("ID del escalador a eliminar | ID: ");
                         int idEliminar = scanner.nextInt();
-                        escaladorDAO.deleteTable(idEliminar);
-                        System.out.println("Escalador eliminado con éxito.");
+                        if (!doesEscaladorExist(idEliminar)) {
+                            System.err.println("El escalador no existe.");
+                            break;
+                        } else {
+                            escaladorDAO.deleteTable(idEliminar);
+                            System.out.println("Escalador eliminado con éxito.");
+                        }
                         break;
-
                     case 5:
+                        System.out.println("--- Mostrar todos los Escaladores ---");
+                        // Mostrar tots els escaladors
+                        escaladorDAO.readAll();
+                        break;
+                    case 6:
                         System.out.println("--- Salir ---");
                         // Exit
                         System.out.println("Saliendo del menú...");
@@ -135,16 +162,126 @@ public class ProvaEscalador {
         }
     }
 
-    public static String UptadeEstilPreferit() {
+    // CONTROLADOR DE EXCEPTIONS
+    public static int insertID () {
+        int id = 0;
+        try {
+            id = scanner.nextInt();
+            if (id < 0) {
+                throw new NegativeNumberException("El ID no pot ser negatiu.");
+            }
+        } catch (NegativeNumberException e) {
+            System.err.println("Error: " + e.getMessage());
+        }
+        return id;
+    }
+
+    public static String insertName () {
+        String name = null;
+        try {
+            name = scanner.nextLine();
+            if (name.isEmpty()) {
+                throw new IllegalArgumentException("El nom no pot estar buit.");
+            }
+        } catch (IllegalArgumentException e) {
+            System.err.println("Error: " + e.getMessage());
+        }
+        return name;
+    }
+
+    public static String insertAlias () {
+        String alias = null;
+        try {
+            alias = scanner.nextLine();
+            if (alias.isEmpty()) {
+                throw new IllegalArgumentException("L'alias no pot estar buit.");
+            }
+        } catch (IllegalArgumentException e) {
+            System.err.println("Error: " + e.getMessage());
+        }
+        return alias;
+    }
+
+    public static int insertAge () {
+        int age = 0;
+        try {
+            age = scanner.nextInt();
+            if (age < 0) {
+                throw new EdatException("L'edat no pot ser negativa.");
+            } else if (age > 120) {
+                throw new EdatException("L'edat no pot ser superior a 120 anys.");
+            }
+        } catch (EdatException e) {
+            System.err.println("Error: " + e.getMessage());
+        }
+        return age;
+    }
+
+    public static String insertNivellMaxim () {
+        String nivellMaxim = null;
+        String regEx = "^(4\\+?|5\\+?|6[abc]\\+?|7[abc]\\+?|8[abc]\\+?|9[abc]\\+?)$";
+        try {
+            nivellMaxim = scanner.nextLine();
+            if (nivellMaxim.matches(regEx)) {
+                throw new GrauDificultatException("El nivell no està dins del rang vàlid.");
+            }
+        } catch (GrauDificultatException e) {
+            System.err.println("Error: " + e.getMessage());
+        }
+        return nivellMaxim;
+    }
+
+    public static String insertEstilPreferit() {
         String nuevoEstilo = null;
         try {
             nuevoEstilo = scanner.nextLine();
             if (!nuevoEstilo.equalsIgnoreCase("esportiva") && !nuevoEstilo.equalsIgnoreCase("clàssica") && !nuevoEstilo.equalsIgnoreCase("gel")) {
-                throw new personalExceptions.EstilPreferitException("Estil preferit no vàlid, ha de ser; esportiva | clàssica | gel");
+                throw new EstilPreferitException("Estil preferit no vàlid, ha de ser; esportiva | clàssica | gel");
             }
-        } catch (personalExceptions.EstilPreferitException e) {
+        } catch (EstilPreferitException e) {
             System.err.println("Error: " + e.getMessage());
         }
         return nuevoEstilo;
+    }
+
+    public static boolean doesViaExist(int idVia, Connection connection) {
+    String sql = "SELECT COUNT(*) FROM vies WHERE id_via = ?";
+        try (PreparedStatement stmt = connection.prepareStatement(sql)) {
+        stmt.setInt(1, idVia);
+        ResultSet rs = stmt.executeQuery();
+        if (rs.next()) {
+            return rs.getInt(1) > 0;
+        }
+    } catch (SQLException e) {
+        e.printStackTrace();
+    }
+    return false;
+    }
+
+    public static int insertViaMax () {
+        int idViaMax = 0;
+        try {
+            idViaMax = scanner.nextInt();
+            if (!doesViaExist(idViaMax, connection)) {
+                throw new IllegalArgumentException("La via no existeix.");
+            }
+        } catch (IllegalArgumentException e) {
+            System.err.println("Error: " + e.getMessage());
+        }
+        return idViaMax;
+    }
+
+    public static boolean doesEscaladorExist (int idEscalador) {
+        String sql = "SELECT COUNT(*) FROM escaladors WHERE id_escalador = ?";
+        try (PreparedStatement stmt = connection.prepareStatement(sql)) {
+            stmt.setInt(1, idEscalador);
+            ResultSet rs = stmt.executeQuery();
+            if (rs.next()) {
+                return rs.getInt(1) > 0;
+            }
+        } catch (SQLException e) {
+            e.printStackTrace();
+        }
+        return false;
     }
 }
